Initial commit: Added basic index.html for GitHub Pages
This is the first commit of the repository. It includes a simple HTML file (index.html) that serves as the homepage for the GitHub Pages site.

The page includes a basic structure with a title, heading, and a welcome message. This will allow the site to be published at https://kanhaminz.github.io/krushnakant/ using GitHub Pages.

Further improvements will include styling, additional sections, and content updates.